feat: 달력 하단에 해당 월의 총 수입, 총 지출, 총계 표시
- getTotalMoney 함수 분리
- getTotalMoney 함수 결과에 sum 프로퍼티 추가
